Created infinite spawner prefab
This prefab will spawn an object when the player enters the box collider.  It will keep spawning objects until the player leaves the collider.  It can be turned off if the infiniteSpawnOn variable in the game controller is set to false.
The box colider can be resized and the pos1 box collider in the prefab can be moved and is the point in which the object will spawn.