Added AOE damage to mines
The mine has a new variable AOEDamageAmount. When the mine explodes everything in the new DamageRadius collider will take that amount of damage. The mine tracks everything that goes into the radius. If there is an object in the scene before the games starts that you want to do damage to, it needs to be added to the collider variable in the datamageRadius object. The test case was having another near by mine explode when the mine goes off which makes it also explode.