Web GL Build and itch.io deploy
Built a webgl version of the game.  Changed build setting to make it run.
set override for for compression to get under the 200mb limit for web.data that is required by itch.io
setup itch.io account and deployed the game
https://projectshadowgame.itch.io/project-shadow
the password is shadow

I also looked at deploying it as a static web app on azure, but it needed a deployment pipline which put it out of scope.